Step by Step Method_CSV and Time Iterations
Any value that is required to be input has been recorded in the CSV file and is fetched through the code. The code to get tao through time iterations has been coded, but has been commented out and the current tao from [28,100,30000] has been used. On the CSV file for things like A_S, each column corresponds to a layer of strands.