Updated 5.10_Fixing
Deflection Calculated - the point along the beam for which the deflected values are calculated is in line 1049 - this is a user input. Also for the calculation of t, the deflected values at the loading points seem okay, but when I add the increments, the graph looks a bit funny. Not sure if this is how it is supposed to be. The first integrated graph versus length and the second integrated graph versus length are both parabolic and make sense - in regards to the deflection along a beam.