This is the Final Code with no hardcoding
Note that user inputs for L and point along beam (eg if they want a time versus deflection for the deflection at L=5m) to plot graph needs to be added, if we decide to. Changes to this are that effects of the integration are mitigated but limiting the plotting range and the t_comp is 68>t_slab which is 40.